requirements.txt
asgiref==3.5.0
bubbling-firebase-authentication @ git+https://github.com/dimsamaras/bubbling_firebase_authentication.git@c42a7aadecc03d89f5f30cc4c5dfb1d3e878c73b
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-split-settings==1.1.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.0
ecdsa==0.17.0
filelock==3.4.2
firebase-admin==5.2.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
identify==2.4.6
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
nodeenv==1.6.0
packaging==21.3
platformdirs==2.4.1
pre-commit==2.17.0
proto-plus==1.19.9
protobuf==3.19.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dotenv==0.19.2
python-jose==3.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
soft-delete @ git+https://github.com/dimsamaras/soft-delete.git@7dca7cfbf83a496bdf681e8ea37337a5ca05482a
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
